All files relating to the app.
committing all files relating to the streamlit app, this commit contains the .h5 file of the bad generator. while training the discriminator became too strong and the generator couldn't learn much. It's worth learning from such mistakes so I've added this in the first commit, in the following ones a proper working generator will replace this one